fix: cjs -> js
