+++ eslint.config.js
@@ -26,6 +26,7 @@ export default tseslint.config(
       "@typescript-eslint/no-explicit-any": "off",
     },
   },
+
   {
     files: ["src/**/*.ts", "src/**/*.tsx"],
     languageOptions: {
